+++ glide-core/src/request_type.rs
@@ -143,6 +143,7 @@ pub enum RequestType {
     Touch = 132,
     ZRevRank = 133,
     ZInterStore = 134,
+    HRandField = 135,
 }
 
 fn get_two_word_command(first: &str, second: &str) -> Cmd {
@@ -289,6 +290,7 @@ impl From<::protobuf::EnumOrUnknown<ProtobufRequestType>> for RequestType {
             ProtobufRequestType::Touch => RequestType::Touch,
             ProtobufRequestType::ZRevRank => RequestType::ZRevRank,
             ProtobufRequestType::ZInterStore => RequestType::ZInterStore,
+            ProtobufRequestType::HRandField => RequestType::HRandField,
         }
     }
 }
@@ -431,6 +433,7 @@ impl RequestType {
             RequestType::Touch => Some(cmd("TOUCH")),
             RequestType::ZRevRank => Some(cmd("ZREVRANK")),
             RequestType::ZInterStore => Some(cmd("ZINTERSTORE")),
+            RequestType::HRandField => Some(cmd("HRANDFIELD")),
         }
     }
 }